Create a frontend application that will show the ship positions on a map

+++ decoder/src/decoders/position_report_class_a.rs
@@ -71,7 +71,7 @@ pub fn get(bytestring: &str) -> Result<MessageData, NMEADecoderError> {
     let true_heading = match get_unsigned_number(&bytestring[128..137]) {
         Ok(n) => match n {
             511 => None,
-            _ => Some(n as u8),
+            _ => Some(n as u16),
         },
         Err(e) => return Err(e),
     };
